refactor(openai-session): split the CLI out so the promoted entrypoint fits
Carrying #1032's entrypoint onto develop pushed openai_session.py to 666 lines
against this repo's 512-line cap. Splitting is the cap's own remedy and the
right shape anyway: the module is a session LIBRARY, and argument parsing plus
process lifecycle are a separate responsibility.

_parse_argv and main move to _runners/_openai_session_cli.py and are
re-exported, so `python -m scitex_agent_container._runners.openai_session` —
the module name the apptainer argv builder emits — still resolves, as does
every existing import.

Two reconciliations the cherry-pick required, both because main and develop
had drifted apart:

  * `__all__` conflicted because each branch appended different names. Both
    sets belong; neither side was wrong.
  * main() constructed `OpenAISession`. On develop the class is
    `OpenAIAgentsSession` — same object, renamed. That rename is why the two
    branches could not be merged wholesale (43 conflicting files), and why
    this lands as two cherry-picks instead.

Verified: OpenAIAgentsSession, OpenAISessionError, main, _parse_argv,
build_mcp_server and McpConfigError all import, and _openai_mcp,
runtimes.openai_session and _store_plugin now coexist on one branch for the
first time.
